requirements.txt
langchain
langchain-community
langchain-core
langchain-tools
langchain-groq
fastapi
uvicorn
pydantic
python-dotenv